pyproject.toml
[tool.poetry]
name = "model-trainer"
version = "0.1.0"
description = ""
authors = ["Adam Yodinsky <adam.yodinsky@autodesk.com>"]
readme = "README.md"
packages = [{include = "model_trainer"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "^1.23.4"
pandas = "^1.5.1"
sklearn = "^0.0"
keras = "^2.10.0"
munch = "^2.5.0"
pyyaml = "^6.0"
matplotlib = "^3.6.1"
psycopg2-binary = "^2.9.5"
tensorflow-macos = "^2.10.0"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
